+++ example/app/app.component.e2e.spec.ts
@@ -6,8 +6,4 @@ describe("AppComponent", () => {
         browser.get("/");
     });
 
-    it("should have a title", () => {
-        expect(browser.getTitle()).toEqual("ng2 Dynamics Forms Sample App");
-    });
-    
 });