+++ modules/ui-basic/src/dynamic-form-basic.component.spec.ts
@@ -41,7 +41,7 @@ describe("DynamicFormBasicComponent test suite", () => {
 
         formGroup = service.createFormGroup(formModel);
 
-        component.controlGroup = formGroup;
+        component.group = formGroup;
         component.model = formModel[0];
 
         fixture.detectChanges();
@@ -52,7 +52,7 @@ describe("DynamicFormBasicComponent test suite", () => {
     it("should initialize correctly", () => {
 
         expect(component.control instanceof FormControl).toBe(true);
-        expect(component.controlGroup instanceof FormGroup).toBe(true);
+        expect(component.group instanceof FormGroup).toBe(true);
         expect(component.model instanceof DynamicFormControlModel).toBe(true);
         expect(component.hasErrorMessaging).toBe(false);
 
@@ -71,7 +71,7 @@ describe("DynamicFormBasicComponent test suite", () => {
         expect(component.isInvalid).toBe(false);
         expect(component.showErrorMessages).toBe(false);
 
-        expect(component.controlType).toBe(CoreFormControlType.Input);
+        expect(component.type).toBe(CoreFormControlType.Input);
     });
 
     it("should have an input element", () => {